+++ src/index.ts
@@ -1,3 +1,3 @@
 export * from './layout';
 export * from './components';
-export * from './functions/getCssValue';
+export * from './utils/getCssValue';
